demo_virtualshop_orders/requirements.txt
google-cloud-pubsub
psutil